[RED] Adding a test that check a simple discount with two books

+++ katapotter/src/main/java/com/flipper83/katapotter/BasketCalculator.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class BasketCalculator {
   private static final int BOOK_PRICE = 8;
 
-  public int calculatePrice(List<String> books) {
+  public float calculatePrice(List<String> books) {
     int amount = BOOK_PRICE * books.size();
     return amount;
   }
